MCPi Info + EntitiesDat Parsing
I initially set this concept up with my experimentation a few days ago, on the 5th.

Looking back at my messages/notes on this, I forgot that I was originally going to look into the format that classic.minecraft.net uses.

Here's the notes on that (these were from the night of the 4th):

--start--

I'm curious about looking into the world format that classic.minecraft.net uses as well, it seems to just be a JSON file with the modified blocks and world seed in it
You can load the same world across sessions/browsers through Local Storage

So I'm curious how this could be utilized to either make custom maps for classic.minecraft.net, or to be able to play old-alpha worlds there as well, because it seems like this version was rewritten/based off of 0.0.23a_01
https://www.reddit.com/r/gaming/comments/bm4zlq/minecraft_classic_is_now_available_to_play_in/

If I can get a version of 0.0.23a_01 running as well, I'll look into the world's NBT format to see if it's similar to how this JSON file works as well
If so, I can probably hack in support to my project Minefork, so you can just use plain old Alpha worlds right in-game haha
Did that version have game saving actually? I'm worried it won't when I am about to go look into it haha

f
https://minecraft.wiki/w/Classic_0.0.23a_01_(remake)#Differences_and_inaccuracies_from_the_original
https://minecraft.wiki/w/Java_Edition_Classic_0.0.23a_01

Well, ended up going a different route haha (12 AM on the 5th, this is in reference to a screenshot I took of me playing MCPI Revival)
https://github.com/MCPI-Revival/minecraft-pi-reborn

Interesting! Since Minecraft Pi was based on old PE versions, that level header for the level.dat was still around back then as well 🕵️ (I started looking into the file format for the files from the savegame of that version, it's the ones I am using for this same repo)

NBTify works in the CLI on the Pi too 🔥 (Managed to do `npm install -g nbtify` on the Raspberry Pi, and it works great for reading NBT files!)

%Bigjango 👋 woah wait didn't know you are a contributor to Pi Edition Reborn!

Looking into the rest of the files as well
Seems that the chunks.dat is a binary format, not NBT, and entities.dat is an NBT file, but it also has a custom header like the level.dat does
https://minecraft.wiki/w/Bedrock_Edition_level_format/History#entities.dat (This is the same experimentation for the resulting code present with this commit)

Maybe I should add support for other header-based files, do you think that is in the scope of NBTify, or not specifically?
I thought the level.dat was the only file that had a unique header like that, so now I have to add custom handling for the entities.dat file too
So that could either be part of Bedrock-LevelDB (Or more specifically, the Bedrock binding, once that has a better name), or it can be part of NBTify, so it can open this file on it's own as well, like the level.dat can be
Because right now the entities.dat can't be opened with plain NBTify or Dovetail because of the header

Wrote a little script to figure out parsing it (the demo for this same code for the entities.dat)
```js
// @ts-check

import { readFile } from "node:fs/promises";
import { inspect } from "node:util";
import * as NBT from "/opt/homebrew/lib/node_modules/nbtify/dist/index.js";

const demo = new URL("./entities.dat",import.meta.url);

const buffer = await readFile(demo);
console.log(buffer.subarray(0,12));

const header = {
  type: buffer.subarray(0,3).toString(),
  a: buffer.readUInt8(3),
  b: buffer.readUInt32LE(4),
  byteLength: buffer.readUInt32LE(8)
};
console.log(header);

/** @type {Required<NBT.ReadOptions>} */
const format = {
  rootName: true,
  endian: "little",
  compression: null,
  bedrockLevel: null,
  strict: true
};

const nbt = await NBT.read(buffer.subarray(12),format);
console.log(inspect(nbt,{ colors: true, depth: null }));

// https://github.com/MCPI-Revival/minecraft-pi-reborn
// https://minecraft.wiki/w/Bedrock_Edition_level_format/History#entities.dat
// The file has a 12 byte header. It begins with the ASCII Characters "ENT", then one zero byte, then a little-endian integer with the value 1, followed by another little endian integer stating the length of the file in bytes, not counting the header.
```

(This then lead to me writing the typings for this version in Region-Types)
https://github.com/Offroaders123/Region-Types/commit/a6c0bc96ed11d381f50004e1025062c8480ca5ea

(The links useful in that commit that are related to this project more specifically)
https://github.com/MCPI-Revival/pi_world
https://github.com/MCPI-Revival/MCPIedit
https://mcpirevival.miraheze.org/wiki/MCPI_Revival
https://github.com/ReMinecraftPE/mcpe

https://minecraft.wiki/w/Bedrock_Edition_level_format/History#0.8.1_and_below (I think this works the same for this one as well)

(Then I worked a bit more on the types for this edition, and had some other useful links to show)
https://github.com/Offroaders123/Region-Types/commit/8c5999c7f5b1faf31e00970e8665af496e7902e1

https://mcpirevival.miraheze.org/wiki/Minecraft:_Pi_Edition_v0.1.1
https://minecraft.wiki/w/Bedrock_Edition_level_format/History#0.8.1_and_below
https://minecraft.wiki/w/Bedrock_Edition_data_values#Entity_IDs
https://minecraft.wiki/w/Bedrock_Edition_level_format/v0.1.0_alpha (inspiration for what I want to do for LCE's docs on the wiki eventually)
https://www.reddit.com/r/technicalminecraft/comments/ozk8ul/does_anybody_know_where_i_can_find_the_source/ (I want to look at the code for older versions, to learn about the shapes for various NBT features)
https://minecraft.wiki/w/Tutorials/Programs_and_editors/Mod_Coder_Pack#History

(There were some extra bonus fun links from those commits too, but I left those out of this list to keep things specific to this project since it's already a lot of things to read-up on haha)

--end--