Chunk Parser: Fixed-Length Chunk List
This is brought over from MCA-Buffer as well, the concept that an array can be used to store a set of chunks, by way of using the array's length to discern what a given chunk's coordinate is. This setup also only allows you to put in the amount of chunks that a world can handle. If you try to push to an array of chunks, it shouldn't let you put in a coordinate that isn't saveable into the world, hence why it is a sealed array.