Chunk Parser: Chunks vs Entries
Making this distinction again, similar to the concept with the Region parser. There's the data entries for the archive/save file itself, then there is the parsed-out structure of the chunks themselves. I don't want to have to require you to parse out all of the chunks at once just to get to a specific part, each stage of reading should be when you need to use it.