Chunk Parser: Entry Property Naming
This is now consistent with my Region handler. Trying to keep the same naming consistencies for each implementation helps to find similarities in logic and formatting for each version. Eventually having a single platform to work with worlds will be easier to accomplish once all of the similarities and differences between each platform can be worked out, and described. It also not only helps to have this in a code-state too, since seeing how code has changed seems to be easier than how a set of documentation for how the file format has changed. So far it seems like the wrapper for describing these chunks in the file is essentially the same as a Region file, except with values being little-endian, a lack of a timestamp section, and the use of NBT. Disregarding the values for Chunk entries (the NBT part), the Region wrapper format thing is fairly consistent for each one. I'm also curious about what kinds of validation I can write to auto-detect which world format the given world or file uses, similar to what NBTify does. Right now, I'm thinking using an enum of config values for each of the types, and checking whether those are the correct ones used throughout the file. Say like in Java how there are a known set of compression formats. If all of the Chunk entries don't use one of the valid ones, you'd know it likely isn't a Java Region file. Those kind of deduction calculations. If one thing can't work, then the rest couldn't either. That's how NBTify works. If you can't safely parse the file all the way through in a given format, then it's highly certain to not be in that format. The file is either correctly stored, or it's not. And if it doesn't quite match the spec you wrote the program for, then it's a new case to account for, in a new separate check.

Going to move the plain entry `length` calculation value into the entry handling, which will be a set of separate functions from the entry discovery process.