Chunk Parser: Chunk Types
Setting up the parsing for each of the sections for a given Chunk. Note that this is not NBT, but I am using NBT primitives because it will make things simpler to understand for what these value equate to in the world formats that do use NBT file saving.

Now I am interested again about looking into the Alpha Java format, as once again it does look similar to this one in certain aspects.

I think an interesting part here is that none of the data is compressed, nor paletted.