Chunk Parser: Sections Offsets
Now I am reading the blocks for each part of a chunk. I double-checked that these values indeed should be correct (they seem to look correct as well), but I'm still a bit skeptical in my math to calculate the offsets for everything. These values are the hex counterparts to the Wiki's description of the offsets.