Chunk Parser: Region Types IDs
Using Region-Types to figure out if the data I'm reading is correct or not.

Also going to use the raw data for the data and light chunk values so I can get a better understanding as for what their underlying meaning are. It's mentioned in the wiki, but I don't fully conceptualize how it works just yet.