LevelDat Types + BedrockLevel Header!!!
Just had an *aha* moment! The version number byte at the start of Bedrock's `level.dat` file is the same as the `StorageVersion` NBT key itself!
So this is something that can be actively typed-out in terms of a discriminated union type as well!

Essentially,
```ts
if (level.StorageVersion.valueOf() === 3){
  level; // old
} else {
  level; // new
}
```

I thought the only way to get that info was through the file header itself, but maybe we can instead derive that from the NBT itself! This is extra cool.
So for NBT handling, maybe `bedrockLevel` can be a boolean instead! And it will instead only look for whether a `StorageVersion` key is present. And if one is not, then it can error or something.
Then you only have a single definition (or, source of truth) for handling that.

https://minecraft.wiki/w/Bedrock_Edition_level_format/History#NBT_Structure

Cornbread Crumb, Kidzapunk