Region-Types Module Installation
This hasn't worked previously before this, and I couldn't quite figure out why. I think it was because I wasn't properly building it for use in a JS/npm context, in that you have to make sure to build it, so then the `package.json` can properly resolve from the `dist` folder. Aah wait, I think it's because I was installing from GitHub, but that makes it so the `src` folder in Region-Types can't be built from. Using Git Submodules mitigates this because it allows you to properly build from source locally in your clone of this project.

https://www.youtube.com/watch?v=nV5kP19YkFM
https://www.youtube.com/watch?v=bEkOuLOEMV8